basicwidget: Text.TextSize should take the container size
Now TextInput uses Text.TextSize to set the size of its text, but
TextSize used its actual size as the outside size, and this is
semantically incorrect: the text width was used as the outside width
next time, and this is an infinite recursive.

This change fixes this issue by adding an argument for TextSize
to take a container width explicitly.